+++ docs/apidocs/edu/ucsb/cs56/ucsb_courses_search/MySearchResult.html
@@ -125,7 +125,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/lang/Object.html?is-external=true" title="class or interface in java.lang" class="externalLink">java.lang.Object</a></li>
 <li>
 <ul class="inheritance">
-<li>edu.ucsb.cs56.ucsb_courses_search.MySearchResult</li>
+<li>edu.ucsb.cs56.ucsb_courses_search.model.result.MySearchResult</li>
 </ul>
 </li>
 </ul>
