+++ docs/apidocs/edu/ucsb/cs56/ucsb_courses_search/class-use/MySearchResult.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (11.0.4) on Wed Nov 06 10:41:20 PST 2019 -->
-<title>Uses of Class edu.ucsb.cs56.ucsb_courses_search.MySearchResult (ucsb_courses_search 0.1.0 API)</title>
+<title>Uses of Class edu.ucsb.cs56.ucsb_courses_search.model.result.MySearchResult (ucsb_courses_search 0.1.0 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2019-11-06">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
@@ -95,9 +95,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </header>
 <main role="main">
 <div class="header">
-<h2 title="Uses of Class edu.ucsb.cs56.ucsb_courses_search.MySearchResult" class="title">Uses of Class<br>edu.ucsb.cs56.ucsb_courses_search.MySearchResult</h2>
+<h2 title="Uses of Class edu.ucsb.cs56.ucsb_courses_search.model.result.MySearchResult" class="title">Uses of Class<br>edu.ucsb.cs56.ucsb_courses_search.model.result.MySearchResult</h2>
 </div>
-<div class="classUseContainer">No usage of edu.ucsb.cs56.ucsb_courses_search.MySearchResult</div>
+<div class="classUseContainer">No usage of edu.ucsb.cs56.ucsb_courses_search.model.result.MySearchResult</div>
 </main>
 <footer role="contentinfo">
 <nav role="navigation">
