+++ docs/apidocs/edu/ucsb/cs56/ucsb_courses_search/class-use/SearchController.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (11.0.4) on Wed Nov 06 10:41:20 PST 2019 -->
-<title>Uses of Class edu.ucsb.cs56.ucsb_courses_search.SearchController (ucsb_courses_search 0.1.0 API)</title>
+<title>Uses of Class edu.ucsb.cs56.ucsb_courses_search.controller.SearchController (ucsb_courses_search 0.1.0 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2019-11-06">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
@@ -95,9 +95,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </header>
 <main role="main">
 <div class="header">
-<h2 title="Uses of Class edu.ucsb.cs56.ucsb_courses_search.SearchController" class="title">Uses of Class<br>edu.ucsb.cs56.ucsb_courses_search.SearchController</h2>
+<h2 title="Uses of Class edu.ucsb.cs56.ucsb_courses_search.controller.SearchController" class="title">Uses of Class<br>edu.ucsb.cs56.ucsb_courses_search.controller.SearchController</h2>
 </div>
-<div class="classUseContainer">No usage of edu.ucsb.cs56.ucsb_courses_search.SearchController</div>
+<div class="classUseContainer">No usage of edu.ucsb.cs56.ucsb_courses_search.controller.SearchController</div>
 </main>
 <footer role="contentinfo">
 <nav role="navigation">
