+++ src/main/java/edu/ucsb/cs56/ucsb_courses_search/controller/SearchByDeptController.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.ucsb_courses_search.controllers;
+package edu.ucsb.cs56.ucsb_courses_search.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,15 +6,12 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import edu.ucsb.cs56.ucsb_courses_search.CurriculumService;
-import edu.ucsb.cs56.ucsb_courses_search.results.CourseListingRow;
-import edu.ucsb.cs56.ucsb_courses_search.results.CourseOffering;
-import edu.ucsb.cs56.ucsb_courses_search.searches.SearchByDept;
-import edu.ucsb.cs56.ucsbapi.academics.curriculums.v1.classes.Course;
+import edu.ucsb.cs56.ucsb_courses_search.service.CurriculumService;
+import edu.ucsb.cs56.ucsb_courses_search.model.result.CourseListingRow;
+import edu.ucsb.cs56.ucsb_courses_search.model.result.CourseOffering;
+import edu.ucsb.cs56.ucsb_courses_search.model.search.SearchByDept;
 import edu.ucsb.cs56.ucsbapi.academics.curriculums.v1.classes.CoursePage;
-import edu.ucsb.cs56.ucsbapi.academics.curriculums.utilities.Quarter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.slf4j.Logger;
