+++ src/main/java/edu/ucsb/cs56/ucsb_courses_search/downloaders/CoursePageToCSV.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.ucsb_courses_search;
+package edu.ucsb.cs56.ucsb_courses_search.downloaders;
 
 import java.io.PrintWriter;
