+++ src/main/java/edu/ucsb/cs56/ucsb_courses_search/service/CurriculumService.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.ucsb_courses_search;
+package edu.ucsb.cs56.ucsb_courses_search.service;
 
 public interface CurriculumService {
     public String getJSON(String subjectArea, String quarter, String courseLevel);
